+++ src/backend/ci/core/dockerhost/biz-dockerhost/src/main/kotlin/com/tencent/devops/dockerhost/init/BuildClusterCronConfiguration.kt
@@ -26,10 +26,10 @@
 
 package com.tencent.devops.dockerhost.init
 
+import com.tencent.devops.dockerhost.cron.DockerTaskRunner
 import com.tencent.devops.dockerhost.cron.Runner
 import com.tencent.devops.dockerhost.services.DockerHostBuildService
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.beans.factory.annotation.Value
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
@@ -49,29 +49,26 @@ import java.util.concurrent.Executors
 @EnableScheduling
 class BuildClusterCronConfiguration : SchedulingConfigurer {
 
-    @Value("\${dockerCli.clearLocalImageCron:0 0 2 * * ?}")
-    var clearLocalImageCron: String? = null
-
     override fun configureTasks(scheduledTaskRegistrar: ScheduledTaskRegistrar) {
         scheduledTaskRegistrar.setScheduler(Executors.newScheduledThreadPool(100))
         scheduledTaskRegistrar.addFixedRateTask(
             IntervalTask(
-                Runnable { runner.startBuild() }, 5000, 60 * 1000
+                Runnable { dockerTaskRunner.startBuild() }, 5000, 60 * 1000
             )
         )
 
         scheduledTaskRegistrar.addFixedRateTask(
             IntervalTask(
-                Runnable { runner.endBuild() }, 20 * 1000, 120 * 1000
+                Runnable { dockerTaskRunner.endBuild() }, 20 * 1000, 120 * 1000
             )
         )
     }
 
     @Autowired
-    private lateinit var runner: Runner
+    private lateinit var dockerTaskRunner: DockerTaskRunner
 
     @Bean
-    fun runner(dockerHostBuildService: DockerHostBuildService): Runner {
-        return Runner(dockerHostBuildService)
+    fun dockerTaskRunner(dockerHostBuildService: DockerHostBuildService): DockerTaskRunner {
+        return DockerTaskRunner(dockerHostBuildService)
     }
 }
